backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
sentence-transformers==2.2.2
openai==1.3.7
pydantic==2.5.0
python-dotenv==1.0.0